fix: Move broadcastExecutionUpdate to utility file to fix Next.js route export
- Create lib/monitoring/broadcast-utils.ts with broadcastExecutionUpdate function
- Remove invalid export from stream/route.ts
- Import broadcastExecutionUpdate from utility module
- Fixes "broadcastExecutionUpdate is not a valid Route export field" build error
