+++ lib/monitoring/broadcast-utils.ts
@@ -1,0 +1,38 @@
+/**
+ * Broadcast utilities for WebSocket/SSE updates
+ * Not currently used (placeholder for future WebSocket implementation)
+ */
+
+interface ExecutionUpdate {
+  execution_id: string;
+  agent_id: string;
+  status: string;
+  total_tokens: number;
+  total_cost_usd: number;
+  start_time: string;
+  end_time?: string;
+  timestamp: string;
+}
+
+// In-memory client registry (in production, use Redis)
+const clients = new Map<string, Set<WebSocket>>();
+
+/**
+ * Broadcast execution update to all subscribed clients
+ * (Placeholder for future WebSocket implementation)
+ */
+export function broadcastExecutionUpdate(update: ExecutionUpdate) {
+  const agentClients = clients.get(update.agent_id);
+  if (!agentClients) return;
+
+  const message = JSON.stringify(update);
+  agentClients.forEach((ws) => {
+    try {
+      if (ws.readyState === WebSocket.OPEN) {
+        ws.send(message);
+      }
+    } catch (error) {
+      console.error('Failed to send execution update:', error);
+    }
+  });
+}
